requirements.txt
langchain
langchain-core
langchain-community
groq
langchain-groq
chromadb
sentence-transformers
pypdf
pymupdf
Pillow
fastapi
uvicorn
streamlit
python-multipart
requests
pytesseract
opencv-python
